requirements.txt
xmljson
kafka-python